+++ resources/views/pages/company-login.blade.php
@@ -1,5 +1,20 @@
 @extends('layouts.website')
 
+@section('meta')
+<title>Company Login | {{config('app.name')}}</title>
+<meta property="og:title" content="Company Login | {{config('app.name')}}" />
+<meta property="og:site_name" content="Company Login | {{config('app.name')}}" />
+<meta property="og:type" content="website" />
+<meta property="og:image" content="{{ asset('images/icon.png') }}">
+<meta property="og:url" content="{{config('app.url')}}" />
+<meta name="twitter:card" content="summary">
+<meta name="twitter:title" content="Company Login | {{config('app.name')}}">
+<meta name="twitter:image:src" content="{{ asset('images/icon.png') }}">
+<meta name="twitter:image" content="{{ asset('images/icon.png') }}">
+
+
+@endsection
+
 
 @section('content')
 <x-base-plain>